+++ src/main/java/nextstep/subway/line/application/dto/LineResponse.java
@@ -1,29 +1,25 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.line.application.dto;
 
-import nextstep.subway.domain.Station;
+import nextstep.subway.line.domain.Line;
 
 import java.time.LocalDateTime;
 
-public class StationResponse {
+public class LineResponse {
     private Long id;
     private String name;
+    private String color;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
-    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
-        this.id = id;
-        this.name = name;
-        this.createdDate = createdDate;
-        this.modifiedDate = modifiedDate;
+    public LineResponse() {
     }
 
-    public static StationResponse of(Station station) {
-        return new StationResponse(
-                station.getId(),
-                station.getName(),
-                station.getCreatedDate(),
-                station.getModifiedDate()
-        );
+    public LineResponse(Long id, String name, String color, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+        this.createdDate = createdDate;
+        this.modifiedDate = modifiedDate;
     }
 
     public Long getId() {
@@ -34,11 +30,19 @@ public class StationResponse {
         return name;
     }
 
+    public String getColor() {
+        return color;
+    }
+
     public LocalDateTime getCreatedDate() {
         return createdDate;
     }
 
     public LocalDateTime getModifiedDate() {
         return modifiedDate;
+    }
+
+    public static LineResponse of(Line line) {
+        return new LineResponse(line.getId(), line.getName(), line.getColor(), line.getCreatedDate(), line.getModifiedDate());
     }
 }
