+++ src/main/java/nextstep/subway/line/domain/LineRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.line.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
