+++ src/main/java/nextstep/subway/station/application/dto/StationResponse.java
@@ -1,25 +1,29 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.station.application.dto;
 
-import nextstep.subway.domain.Line;
+import nextstep.subway.station.domain.Station;
 
 import java.time.LocalDateTime;
 
-public class LineResponse {
+public class StationResponse {
     private Long id;
     private String name;
-    private String color;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
-    public LineResponse() {
-    }
-
-    public LineResponse(Long id, String name, String color, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
-        this.color = color;
         this.createdDate = createdDate;
         this.modifiedDate = modifiedDate;
+    }
+
+    public static StationResponse of(Station station) {
+        return new StationResponse(
+                station.getId(),
+                station.getName(),
+                station.getCreatedDate(),
+                station.getModifiedDate()
+        );
     }
 
     public Long getId() {
@@ -30,19 +34,11 @@ public class LineResponse {
         return name;
     }
 
-    public String getColor() {
-        return color;
-    }
-
     public LocalDateTime getCreatedDate() {
         return createdDate;
     }
 
     public LocalDateTime getModifiedDate() {
         return modifiedDate;
-    }
-
-    public static LineResponse of(Line line) {
-        return new LineResponse(line.getId(), line.getName(), line.getColor(), line.getCreatedDate(), line.getModifiedDate());
     }
 }
