+++ src/main/java/nextstep/subway/station/domain/StationRepository.java
@@ -1,8 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.station.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
     boolean existsByName(String name);
